app/requirements.txt
flask
flask-mysql
mysql-connector